use systemd's headers and tabsize, add error-triggering options in mock-server
