+++ src/timesync/meson.build
@@ -87,12 +87,10 @@ if conf.get('ENABLE_TIMESYNC_NTS') == 1
                         'link_with' : libshared,
                         'dependencies' : crypto_deps,
                 },
-                libexec_template + {
-                        'name' : 'nts-mock-server',
-                        'sources' : files('test-server.c', 'nts_crypto_openssl.c', 'nts_extfields.c'),
-                        'extract' : timesyncd_extract_sources,
-                        'dependencies' : [libm, crypto_deps],
-                        'install' : false,
+                test_template + {
+                        'sources' : files('test-nts-mockserver.c') + nts_sources,
+                        'dependencies' : crypto_deps,
+                        'type' : 'manual',
                 },
         ]
 endif
